+++ ast/util/free_ast.h
@@ -15,11 +15,6 @@
 void free_ast(struct AST* ast);
 
 //const
-void free_bool_const(struct BoolConst* bc);
-void free_char_const(struct CharConst* cc);
-void free_int_const(struct IntConst* ic);
-void free_hex_const(struct HexConst* hc);
-void free_bin_const(struct BinConst* hc);
 void free_const_value(struct ConstValue* cv);
 void free_string_const(struct StringConst* s);