+++ compiler/main/gen_tac/gen_tac.c
@@ -6,11 +6,11 @@
 int int_value_from_const(struct ConstValue* cv){
 
     switch (cv->kind) {
-        case 1: return (int) cv->ptr.m1_bool_const->value;
-        case 2: return (int) cv->ptr.m2_int_const->value;
-        case 3: return (int) cv->ptr.m3_char_const->value;
-        case 5: return (int) cv->ptr.m5_hex_const->value;
-        case 6: return (int) cv->ptr.m6_bin_const->value;
+        case 1: return (int) cv->ptr.m1_bool_const;
+        case 2: return (int) cv->ptr.m2_int_const;
+        case 3: return (int) cv->ptr.m3_char_const;
+        case 5: return (int) cv->ptr.m5_hex_const;
+        case 6: return (int) cv->ptr.m5_hex_const;
     }
     return -1;
 }