+++ parser/main/astnodes/const/IntConst.h
@@ -2,10 +2,10 @@
 #define INTCONST
 
 #include <stdbool.h>
+#include <inttypes.h>
 
-struct IntConst;
 struct TokenList;
 
-struct IntConst* makeIntConst(struct TokenList* tokens);
+int32_t makeIntConst(struct TokenList* tokens, bool* error);
 
 #endif