+++ parser/test/astnodes/const/BoolConstTest.h
@@ -1,8 +1,0 @@
-#ifndef BOOLCONSTTEST
-#define BOOLCONSTTEST
-
-#include <stdbool.h>
-
-int test_boolconst_parse_bool_constant_node();
-
-#endif