+++ parser/test/astnodes/const/ConstValueTest.h
@@ -1,10 +1,13 @@
-#ifndef CHARCONSTTEST
-#define CHARCONSTTEST
+#ifndef CONST_VALUE_TEST_H
+#define CONST_VALUE_TEST_H
 
 #include <stdbool.h>
+
+int test_boolconst_parse_bool_constant_node();
 
 int test_charconst_parse_char_constant_node();
 
 int test_charconst_parse_char_constant_node_newline();
+
 
 #endif